prevent panics when cleaning up resources that haven't been created.

+++ builder/oracle/classic/step_create_ip_reservation.go
@@ -42,12 +42,16 @@ func (s *stepCreateIPReservation) Run(_ context.Context, state multistep.StateBa
 }
 
 func (s *stepCreateIPReservation) Cleanup(state multistep.StateBag) {
+	ipResName, ok := state.GetOk("ipres_name")
+	if !ok {
+		return
+	}
+
 	ui := state.Get("ui").(packer.Ui)
 	ui.Say("Cleaning up IP reservations...")
 	client := state.Get("client").(*compute.ComputeClient)
 
-	ipResName := state.Get("ipres_name").(string)
-	input := compute.DeleteIPReservationInput{Name: ipResName}
+	input := compute.DeleteIPReservationInput{Name: ipResName.(string)}
 	ipClient := client.IPReservations()
 	err := ipClient.DeleteIPReservation(&input)
 	if err != nil {
